demo: rename builder to playground, fix layout and stacking bugs
Renaming: the page doesn't produce code, so "builder" was the wrong
name; "playground" matches what it actually is, build a menu and try
it right there.

Fixes, found by trying the page for real:
- The demo's new link to the playground sat under #main, which
  covers the whole page to catch a gesture from anywhere; the link
  never received a click. Give it its own stacking position.
- The playground's own accent color was a plain link blue, at odds
  with the demo's monochrome look (see its GitHub corner icon); use
  the same dark gray instead.
- A menu's fixed-size circle could spill past a narrow preview and
  overlap the item list beside it. Contain it, and let the item list
  wrap onto its own row once the preview no longer has room to draw
  the whole circle.
- Item labels were clipped: this page's sans-serif is taller than the
  library's own default at the size menu.css sizes labels for. Draw
  the menu itself in the browser's plain default font, the same as
  the demo page.
