+++ demo/index.html
@@ -36,7 +36,7 @@
         <p>
           Quick draw to select an item without opening a menu (expert mode).
         </p>
-        <p><a href="./builder/">Build your own menu and try it out</a></p>
+        <p><a href="./playground/">Build your own menu and try it out</a></p>
       </div>
     </div>
