+++ demo/playground/index.html
@@ -3,7 +3,7 @@
   <head>
     <meta charset="UTF-8" />
     <meta content="width=device-width" name="viewport" />
-    <title>Marking Menu Builder</title>
+    <title>Marking Menu Playground</title>
     <link href="./style.css" rel="stylesheet" />
     <script type="module" src="./main.ts"></script>
     <meta
@@ -14,7 +14,7 @@
 
   <body>
     <header>
-      <h1>Marking Menu Builder</h1>
+      <h1>Marking Menu Playground</h1>
       <p>
         Add items, set an angle or leave it free, then draw a stroke on the menu
         to see what the recognizer makes of it. The address updates as you go,
