+++ tsconfig.json
@@ -5,6 +5,6 @@
     { "path": "./tsconfig.test.json" },
     { "path": "./tsconfig.node.json" },
     { "path": "./e2e/tsconfig.json" },
-    { "path": "./demo/builder/tsconfig.json" }
+    { "path": "./demo/playground/tsconfig.json" }
   ]
 }
